+++ tests/Core/MidiInputManagerTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using NAudio.Midi;
 using NUnit.Framework;
 using DavidRice.BlishHud.MidiControl.Core;
 
@@ -8,6 +9,56 @@ namespace DavidRice.BlishHud.MidiControl.Tests.Core
     [TestFixture]
     public class MidiInputManagerTests
     {
+        [Test]
+        public void TryConvertToMidiNoteEvent_NoteOnWithVelocity_ReturnsNoteOn()
+        {
+            var noteOn = new NoteOnEvent(0, 1, 60, 100, 0);
+            var result = MidiEventConverter.TryConvertToMidiNoteEvent(noteOn);
+
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result!.Value.NoteNumber, Is.EqualTo(60));
+            Assert.That(result.Value.IsNoteOn, Is.True);
+        }
+
+        [Test]
+        public void TryConvertToMidiNoteEvent_NoteOnWithZeroVelocity_ReturnsNoteOff()
+        {
+            var noteOn = new NoteOnEvent(0, 1, 60, 0, 0);
+            var result = MidiEventConverter.TryConvertToMidiNoteEvent(noteOn);
+
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result!.Value.NoteNumber, Is.EqualTo(60));
+            Assert.That(result.Value.IsNoteOn, Is.False);
+        }
+
+        [Test]
+        public void TryConvertToMidiNoteEvent_NoteOffEvent_ReturnsNoteOff()
+        {
+            var noteOff = new NoteEvent(0, 1, MidiCommandCode.NoteOff, 60, 0);
+            var result = MidiEventConverter.TryConvertToMidiNoteEvent(noteOff);
+
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result!.Value.NoteNumber, Is.EqualTo(60));
+            Assert.That(result.Value.IsNoteOn, Is.False);
+        }
+
+        [Test]
+        public void TryConvertToMidiNoteEvent_KeyAfterTouch_ReturnsNull()
+        {
+            var afterTouch = new NoteEvent(0, 1, MidiCommandCode.KeyAfterTouch, 60, 100);
+            var result = MidiEventConverter.TryConvertToMidiNoteEvent(afterTouch);
+
+            Assert.That(result, Is.Null);
+        }
+
+        [Test]
+        public void TryConvertToMidiNoteEvent_NullEvent_ReturnsNull()
+        {
+            var result = MidiEventConverter.TryConvertToMidiNoteEvent(null);
+
+            Assert.That(result, Is.Null);
+        }
+
         [Test]
         public void EvaluateConnection_TargetIsNull_ReturnsNoAction()
         {